split modern tests
